Speed optimization by ordering the cases blocks in the state switch statement.
The state "STATE_START_DOCUMENT" only appears one, so this state can
move to the bottom of the switch statement. In my test the most used
state is "STATE_IN_STRING", so we move this state at to top of the
switch statement. With this change the parse process is about 15%
faster.