Avoid unnecessary BOM checks
Overhead of method calls can be significant when parsing very large
JSON files. Only checking for the BOM at the first four characters of
a file makes for an easy performance improvement.

The difference in parsing speed is negligible for files like
'tests/ratherBig.json'. When parsing a 50+ megabyte single line of
JSON with lots of string data and using a light-weight listener, this
improvement can cut the parsing time in half.